Create new Jest module
This module provides a place for utility functions and types that reduce the
boiler plate when writing tests with fp-ts types. This adds:

- `mockStateReaderTaskEither` which provides an easier way to create mocks and
spies to check that the Reader and TaskEither parts get called correctly.

- Matches `toHaveRightValueWithResult` and `toHaveRightValueWithState` which
remove the repetition involved with checking that these StateReaderTaskEither
functions return the right result or state.