Add a lazy stream implementation to read values from tasks
This adds a `Stream.lazy` constructor which takes a generator function that
returns either a `Task` or `TaskEither` which evaluates to an `Option`. The
stream will invoke this generator to get more values until that `Option` is
`None`.

Part of this also involved add some small utility functions in `Either` and
`Refinements`.