FIX Stream.separate
This function was previously implemented incorrectly because it tried to read
from the same iterable multiple times. The act of reading the iterable is not a
pure process and can't be assumed to have no side effects.

This re implements separate to create two new streams which both pull from the
same source as needed.

This function is going to let us stream data through various steps in a lambda
and still end up with a collection of which elements succeeded and which ones
failed.